Steganography/KC-Stega/requirements.txt
absl-py==0.10.0
astor==0.8.1
bleach==1.5.0
certifi==2020.6.20
chardet==3.0.4
enum34==1.1.10
gast==0.4.0
gensim==3.8.3
grpcio==1.31.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==1.7.0
jieba==0.42.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mock==4.0.2
numpy==1.14.0
protobuf==3.13.0
requests==2.24.0
scipy==1.5.2
six==1.15.0
smart-open==3.0.0
tensorboard==1.11.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
